+++ client/src/components/NavBar.js
@@ -1,22 +1,24 @@
 import React from 'react';
+import Select from 'react-select';
 import styled from 'styled-components';
 import { media } from '../styled';
-import Select from 'react-select';
-//import 'react-select/dist/react-select.css';
 
 export default ({ value, onServiceChange }) => (
   <Wrapper>
-    <Logo />
+    <a href="/">
+      <Logo />
+    </a>
     <Title>
       <Select
         className="ServiceSelect"
-        value={value}
         clearable={false}
         options={[
           { value: 'english', label: 'English Sunday Service Roster' },
           { value: 'chinese', label: '中文堂服事表' }
         ]}
         onChange={onServiceChange}
+        searchable={false}
+        value={value}
       />
     </Title>
     <Org>EFC Sydney</Org>
@@ -42,8 +44,8 @@ const Wrapper = styled.div`
 `;
 const Logo = styled.img.attrs({ src: '/logo.png' })`
   display: inline-block;
-  vertical-align: middle;
   height: 32px;
+  vertical-align: middle;
   width: 32px;
 `;
 const Title = styled.span`
@@ -52,7 +54,6 @@ const Title = styled.span`
   font-weight: bold;
   position: relative;
   z-index: 3;
-  min-width: 300px;
   &:link,
   &:hover,
   &:visited {
@@ -60,12 +61,9 @@ const Title = styled.span`
     text-decoration: none;
   }
   ${media.mobile`
-    min-width: 220px;
-  `};
-  ${media.mobile`
     font-size: 14px;
-    position: static;
     left:auto;
+    position: static;
   `};
 `;
 const Org = styled.span`
