+++ client/src/index.css
@@ -14,14 +14,27 @@ body {
 }
 
 /* Title Service Dropdown */
+.ServiceSelect .Select-control {
+  width: auto;
+}
+.ServiceSelect.Select:focus {
+  outline: none;
+}
 .ServiceSelect.Select .Select-control,
 .ServiceSelect.Select.is-focused .Select-control {
   border: none;
   background: transparent;
+  box-shadow: none;
+}
+.ServiceSelect .Select-control .Select-input:focus {
+  background: none;
 }
 .ServiceSelect.Select .Select-arrow,
 .ServiceSelect.Select.is-focused .Select-arrow {
   border-color: #fff transparent transparent;
+}
+.ServiceSelect.Select.is-focused.is-open .Select-arrow {
+  border-color: transparent transparent #999;
 }
 .ServiceSelect.Select.Select--single.is-focused > .Select-control .Select-value .Select-value-label,
 .ServiceSelect.Select.Select--single > .Select-control .Select-value .Select-value-label {
@@ -31,14 +44,15 @@ body {
   background: #fff;
   border-color: #b3b3b3 #ccc #d9d9d9;
 }
-/*
-.ServiceSelect.Select.is-open.Select--single > .Select-control .Select-value {
-  position: static;
-}
-*/
 .ServiceSelect.Select.is-open.Select--single > .Select-control .Select-value .Select-value-label {
   color: #333;
 }
 .ServiceSelect .Select-arrow-zone:hover > .Select-arrow {
   border-top-color: #fff;
+}
+.ServiceSelect.Select.Select--single > .Select-control .Select-value {
+  position: static;
+}
+.ServiceSelect .Select-input {
+  display: none !important;
 }
